Split tests into PropertySyntax tests and TypeSyntax tests
