+++ Sources/SwiftInspectorVisitors/Tests/TypeSyntaxVisitorSpec.swift
@@ -331,303 +331,6 @@ final class TypeSyntaxVisitorSpec: QuickSpec {
         }
       }
 
-      context("when there are multiple properties on the same line") {
-        let content = """
-        public final class FakeType {
-          public var thing: String, foo: Int
-        }
-        """
-
-        it("detects the properties") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          let expectedPropSet: Set<PropertyInfo> = [
-            .init(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.public, .instance]),
-            .init(
-              name: "foo",
-              typeAnnotation: "Int",
-              comment: "",
-              modifiers: [.public, .instance])
-          ]
-          expect(sut.propertiesData) == expectedPropSet
-        }
-      }
-
-      context("when there is a static property") {
-        let content = """
-        public final class FakeType {
-          public static var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.public, .static])
-          ]
-        }
-      }
-
-      context("when there is a static property in reverse order") {
-        let content = """
-        public final class FakeType {
-          static public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.public, .static])
-          ]
-        }
-      }
-
-      context("when there is a private property") {
-        let content = """
-        public final class FakeType {
-          private static var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.private, .static])
-          ]
-        }
-      }
-
-      context("when there is a public private(set) property") {
-        let content = """
-        public final class FakeType {
-          public private(set) var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.public, .privateSet, .instance])
-          ]
-        }
-      }
-
-      context("when there is a public internal(set) property") {
-        let content = """
-        public final class FakeType {
-          public internal(set) var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.public, .internalSet, .instance])
-          ]
-        }
-      }
-
-      context("when there is a internal public(set) property") {
-        let content = """
-        public final class FakeType {
-          internal public(set) var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.publicSet, .internal, .instance])
-          ]
-        }
-      }
-
-      context("when there is a fileprivate property") {
-        let content = """
-        public final class FakeType {
-          fileprivate static var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "",
-              modifiers: [.fileprivate, .static])
-          ]
-        }
-      }
-
-      context("when there is no type annotation") {
-        let content = """
-        public final class FakeType {
-          private static var thing = "Hello, World"
-        }
-        """
-
-        it("detects the property with no type information") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: nil,
-              comment: "",
-              modifiers: [.private, .static])
-          ]
-        }
-      }
-
-      context("when there is a property attribute") {
-        let content = """
-        public final class FakeType {
-          @objc public var thing = "Hello, World"
-        }
-        """
-
-        it("detects the property with no type information") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: nil,
-              comment: "",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
-      context("when there is a line comment") {
-        let content = """
-        public final class FakeType {
-          // The thing
-          public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property with the comment") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "// The thing",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
-      context("when there is a multi-line comment") {
-        let content = """
-        public final class FakeType {
-          // The thing
-          // is such a thing
-          public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property with the comment") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "// The thing\n// is such a thing",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
-      context("when there is a doc line comment") {
-        let content = """
-        public final class FakeType {
-          /// The thing
-          public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property with the comment") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "/// The thing",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
-      context("when there is a block comment") {
-        let content = """
-        public final class FakeType {
-          /* The thing */
-          public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property with the comment") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "/* The thing */",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
-      context("when there is a doc block comment") {
-        let content = """
-        public final class FakeType {
-          /** The thing */
-          public var thing: String = "Hello, World"
-        }
-        """
-
-        it("detects the property with the comment") {
-          try VisitorExecutor.walkVisitor(sut, overContent: content)
-          expect(sut.propertiesData) == [
-            PropertyInfo(
-              name: "thing",
-              typeAnnotation: "String",
-              comment: "/** The thing */",
-              modifiers: [.public, .instance])
-          ]
-        }
-      }
-
     }
   }
 }
